+++ hiring-portal/Backend/routes/otherApiRoutes.js
@@ -1,0 +1,10 @@
+const express=require('express');
+const router=express.Router();
+const otherApiController=require('../controllers/otherApiController');
+
+router.post('/test',otherApiController.test);
+router.post('/hire',otherApiController.hire);
+router.post('/reject',otherApiController.reject);
+router.get('/results/:assessmentId',otherApiController.getResult);
+
+module.exports=router;
